requirements.txt
# Core database drivers
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymysql>=1.0.0
pyodbc>=4.0.0

# NoSQL databases
pymongo>=3.12.0
redis>=4.0.0
json5>=0.9.10

# Data processing and analysis
pandas>=1.3.0
numpy>=1.21.0

# Visualization (static plots only)
matplotlib>=3.4.0
seaborn>=0.11.0

# Logging
loguru>=0.5.3

# Configuration
python-dotenv>=0.19.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0